# use vtproto in bazel_request.go